speed/flink/requirements.txt
apache-flink==2.2.0
influxdb-client
cassandra-driver